requirements.txt
fastapi
uvicorn
motor
beanie
pydantic
pydantic-settings
python-jose[cryptography]
passlib
bcrypt==3.2.0
python-multipart
email-validator
stripe
python-dotenv
